ZOB-371 Fix issues from arch lint rules, add PR workflow
