--- lib/features/appConfiguration/entity/api_configuration.dart
@@ -1,5 +1,0 @@
-enum ApiConfiguration {
-  dev,
-  stage,
-  prod
-}
