--- lib/features/appConfiguration/entity/build_configuration.dart
@@ -1,5 +1,0 @@
-enum BuildConfiguration {
-  dev,
-  stage,
-  prod
-}
