--- lib/features/appConfiguration/mapper/app_configuration_mapper.dart
@@ -1,32 +1,0 @@
-import 'package:zachranobed/features/appConfiguration/entity/api_configuration.dart';
-import 'package:zachranobed/features/appConfiguration/entity/build_configuration.dart';
-
-class AppConfigurationMapper {
-    static BuildConfiguration mapBuildConfiguration(String? flavor) {
-        switch (flavor) {
-            case "dev":
-                return BuildConfiguration.dev;
-            case "stage":
-                return BuildConfiguration.stage;
-            case "prod":
-                return BuildConfiguration.prod;
-            default:
-                // When we can not resolve flavor type, it is safe to use the `prod` option
-                return BuildConfiguration.prod;
-        }
-    }
-
-    static ApiConfiguration mapApiConfiguration(String? flavor) {
-        switch (flavor) {
-            case "dev":
-                return ApiConfiguration.dev;
-            case "stage":
-                return ApiConfiguration.stage;
-            case "prod":
-                return ApiConfiguration.prod;
-            default:
-            // When we can not resolve flavor type, it is safe to use the `prod` option
-                return ApiConfiguration.prod;
-        }
-    }
-}
