+++ lib/features/login/di/login_dependency_container.dart
@@ -1,5 +1,5 @@
 import 'package:get_it/get_it.dart';
-import 'package:zachranobed/features/login/domain/check_if_devtools_are_enabled_usecase.dart';
+import 'package:zachranobed/common/domain/usecase/check_if_devtools_are_enabled_usecase.dart';
 
 class LoginDependencyContainer {
   const LoginDependencyContainer._();
