+++ lib/features/offeredfood/domain/repository/offered_food_repository.dart
@@ -1,4 +1,4 @@
-import 'package:zachranobed/features/foodboxes/domain/model/box_info.dart';
+import 'package:zachranobed/common/domain/model/box_info.dart';
 import 'package:zachranobed/features/offeredfood/domain/model/food_info.dart';
 import 'package:zachranobed/features/offeredfood/domain/model/offered_food.dart';
 import 'package:zachranobed/models/delivery.dart';
