+++ lib/firebase/firebase_options.dart
@@ -1,9 +1,8 @@
 import 'package:firebase_core/firebase_core.dart' show FirebaseOptions;
-import 'package:zachranobed/features/appConfiguration/app_configuration.dart';
-import 'package:zachranobed/features/appConfiguration/entity/api_configuration.dart';
+import 'package:zachranobed/common/domain/model/project_configuration.dart';
 import 'package:zachranobed/firebase/firebase_options_dev.dart' as dev;
-import 'package:zachranobed/firebase/firebase_options_stage.dart' as stage;
 import 'package:zachranobed/firebase/firebase_options_prod.dart' as prod;
+import 'package:zachranobed/firebase/firebase_options_stage.dart' as stage;
 
 /// Default [FirebaseOptions] for use with your Firebase apps.
 ///
@@ -17,7 +16,7 @@ import 'package:zachranobed/firebase/firebase_options_prod.dart' as prod;
 /// ```
 class DefaultFirebaseOptions {
   static FirebaseOptions get currentPlatform {
-    switch (AppConfiguration.instance.apiConfiguration) {
+    switch (ProjectConfiguration.instance.apiConfiguration) {
       case ApiConfiguration.dev:
         return dev.DefaultFirebaseOptions.currentPlatform;
       case ApiConfiguration.stage:
